requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.14.0
biopython>=1.81
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.3
pyyaml>=6.0
tqdm>=4.65.0
wandb>=0.15.0
PyTDC>=1.0
